requirements.txt
django>=6.0,<7.0
python-dotenv>=1.0
pyyaml>=6.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-instrumentation-django>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20
